% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tensor_unfold_cpp}
\alias{tensor_unfold_cpp}
\title{Tensor Unfolding (Matricization) for Synthetic Interventions}
\usage{
tensor_unfold_cpp(T_cube, mode)
}
\arguments{
\item{T_cube}{A 3D array (cube) of dimensions (n1, n2, n3)}

\item{mode}{The mode to unfold along (1, 2, or 3)}
}
\description{
Tensor Unfolding (Matricization) for Synthetic Interventions
}
