% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coresynth-package.R
\docType{package}
\name{coresynth-package}
\alias{coresynth}
\alias{coresynth-package}
\title{coresynth: Fast and Unified Synthetic Control Methods}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

High-performance Synthetic Control Method (SCM) and related causal inference estimators (SDID, GSC, MC, TASC, SI, plus an experimental-design variant) with a unified Formula interface. All computational bottlenecks (QP solving, SVD-based matrix completion, Kalman filtering) are implemented in C++ via RcppArmadillo, providing up to ~55x speed improvements over pure-R alternatives on typical problem sizes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yo5uke/coresynth}
  \item \url{https://yo5uke.com/coresynth/}
  \item Report bugs at \url{https://github.com/yo5uke/coresynth/issues}
}

}
\author{
\strong{Maintainer}: Yosuke Abe \email{yosuke.abe0507@gmail.com}

Authors:
\itemize{
  \item Yosuke Abe \email{yosuke.abe0507@gmail.com}
}

}
\keyword{internal}
