% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.coresynth}
\alias{plot.coresynth}
\title{Plot a coresynth model}
\usage{
\method{plot}{coresynth}(x, type = c("trend", "gap", "weights"), ...)
}
\arguments{
\item{x}{A \code{coresynth} object.}

\item{type}{One of \code{"trend"} (observed vs synthetic), \code{"gap"} (ATT over time),
or \code{"weights"} (donor unit weight bar chart).}

\item{...}{Ignored.}
}
\value{
A \code{ggplot2} plot object.
}
\description{
Plot a coresynth model
}
